requirements.txt
torch>=1.9.0
torchvision>=0.10.0
gradio>=4.0.0
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.21.0
matplotlib>=3.3.0
scikit-image>=0.18.0
